Rework notebooks to monitor error, not coefficients
Monitoring coefficients gets tricky when you move to logistic
regression, where there are an infinite number of solutions when the
data is perfectly separable. Given that, it's probably more appropriate to
monitor the error during training. This also fits more in line with the
way larger neural networks are trained - we monitor error, and by and
large ignore the actual coefficient values.